Move delete method to the message component
