+++ src/components/MessagesList/MessagesGroup/Message/MessageButtonsBar/MessageButtonsBar.vue
@@ -142,8 +142,6 @@ import { generateUrl } from '@nextcloud/router'
 import {
 	showError,
 	showSuccess,
-	showWarning,
-	TOAST_DEFAULT_TIMEOUT,
 } from '@nextcloud/dialogs'
 import Forwarder from '../MessagePart/Forwarder'
 import Button from '@nextcloud/vue/dist/Components/Button'
@@ -374,40 +372,6 @@ export default {
 			EventBus.$emit('focus-chat-input')
 		},
 
-		async handleDelete() {
-			this.isDeleting = true
-			try {
-				const statusCode = await this.$store.dispatch('deleteMessage', {
-					message: {
-						token: this.token,
-						id: this.id,
-					},
-					placeholder: t('spreed', 'Deleting message'),
-				})
-
-				if (statusCode === 202) {
-					showWarning(t('spreed', 'Message deleted successfully, but Matterbridge is configured and the message might already be distributed to other services'), {
-						timeout: TOAST_DEFAULT_TIMEOUT * 2,
-					})
-				} else if (statusCode === 200) {
-					showSuccess(t('spreed', 'Message deleted successfully'))
-				}
-			} catch (e) {
-				if (e?.response?.status === 400) {
-					showError(t('spreed', 'Message could not be deleted because it is too old'))
-				} else if (e?.response?.status === 405) {
-					showError(t('spreed', 'Only normal chat messages can be deleted'))
-				} else {
-					showError(t('spreed', 'An error occurred while deleting the message'))
-					console.error(e)
-				}
-				this.isDeleting = false
-				return
-			}
-
-			this.isDeleting = false
-		},
-
 		handleActionMenuUpdate(type) {
 			if (type === 'open') {
 				this.isActionMenuOpen = true
@@ -459,6 +423,10 @@ export default {
 				console.debug('Current user has already reacted')
 			}
 
+		},
+
+		handleDelete() {
+			this.$emit('delete')
 		},
 	},
 }
